Restructure MockGen into separate parsing and code generation stages.
There is a new subpackage, mockgen/model, which is an abstract
representation of a Go package. The source-based parser constructs a
model.Package, and the code generator works solely from that.
This will make future approaches (e.g. reflection-based parser) easier.

In the process I have cleaned up a bunch of the crufty package-handling
code. This should be easier to work with in the future.