Fix handling of nil passed to EXPECT() methods.
A nil passed to such a method would be wrapped as Eq(nil),
which would only match other nil interface values. That's
not helpful if the method takes a nilable concrete type
(e.g. a pointer). The fix here is to handle nil specially
and store it as Nil() instead.