Fix matching of exhausted calls.
Previously the exhaustion of a call removed the call from the callSet,
but only after the call was matched. If someone used Times(0) then the
call should never match, so we check for exhaustion when searching for
a match and pretend the call doesn't exist if it is exhausted.

Fixes issue 15.