narrow types in Recorder method signatures
This uses the exact types needed in the Recorder methods. It avoids
spurious expectation misses caused by Go converting untyped
constants to
types that do not match the interface generated for.

One example is float literals passed to a `float32` argument were being
converted on x86_64 machines to `float64` because the Recoder methods
had a `interface{}` arg instead of a `float32` arg.

Fixes #16