gomock: fix matcher's degenerate quadratic path.
I noticed several tests timing out when attempting to integrate HEAD
into our mainline branch.  The reason was the accidental introduction of
a quadratic call path in callSet.FindMatch, which performed string
concatenation in a tight loop.

This was corroborated by PProf reporting 23 seconds out of 30 seconds
being spent in runtime.concatstrings, which backs onto runtime.memmove.
The host I instrumented this on was not the same as the ones for which
the data is shown below.

The error accumulation routine has been replaced with bytes.Buffer, which
exponentially grows its underlying buffer to amortize the cost.  After
converting a prototype changelist to this pull request you see here, the
total runtime of this test that demonstrated the defect fell from 80
seconds to 15.  This was just one run on a noisy cluster.

This is change introduced the defect:

https://github.com/golang/mock/blame/4187d4d04aa043124750c9250259ceafdc5f7380/gomock/callset.go#L68