gomock.WithContext: controller with a context which is cancelled on Fatals.
testing.T.FailNow (and Fatal, ...) work only when called from the same
goroutine as the test. This makes using gomock with e.g. RPC server
mocks quite a challenge as the tests are likely to deadlock instead of
failing properly.

This adds a new constructor returning an additional context, which is
cancelled on any fatal failure in mock.