fixed #156 context import collision
This addresses the issue where a package imports another package with context
and golang.org/x/net/context in the same included package. This is leagal since
golang.org/x/net/context is an alias for context. We still don't want other
aliases to context to pass. There are tests that varify that only context and
golang.org/x/net/context are compatable for this special case.