Pass -build_flags as Multiple Args
Value of -build_flags was passed to 'go' as a single argument.
For example when passing -build_flags "-mod vendor", the invoked
commandline was `go -mod\ vendor` and the go command errored out because
there is no such arg.

That has caused confusion, e.g. in #296

Solve this by splitting the argument by a space.