Add -source_package flag.
Currently, mockgen invoked in source mode is relying on x/packages for determining import path of input file.

This behaviour is not symmetrical with -aux_files flag, where import paths are specified explicitly.

This patch adds new optional flag -source_package. When this flag is specified, x/packages is not invoked, and
mock generation becomes completely independent of build system.