mockgen: defer failing on conflicting imports
In source mode, all package files are merged before parsing. This
causes a problem for imports since imports are file scoped. Thus, some
valid packages fail to parse with an error about duplicate imports.

I think that fixing this completely will take some large restructuring
of the parser. However, here I expand the number of packages we can
parse by only raising the duplicate package error if conflicting
package is actually accessed.

For example, we would previously fail to generate a mock for the
following file, but now it succeeds:

    package service

    import "google.golang.org/grpc"

    type Test interface {
    	grpc.ClientStream
    }

Signed-off-by: Steven Danna <steve@chef.io>