gomock/controller: use skip additional frame
In 832173191abead7f0878d8bb24f32f1a5fdf6d79, the callerInfo call
gets nested within an additional function call, but the frame skip
isn't updated. When used with mockgen, this causes the location frame
to unhelpfully point to the generated mock instead of the callsite in
the user's test.

Before:
... at .../mocks/mock_ecriface/mock_ECRAPI.go:612 because: there are no expected calls of...

After:
... at .../example_test.go:29 because: there are no expected calls of...