gomock/controller: use skip additional frame (#443)
In 832173191abead7f0878d8bb24f32f1a5fdf6d79, the callerInfo call
gets nested within an additional function call, but the frame skip
isn't updated. When used with mockgen, this causes the location frame
to unhelpfully point to the generated mock instead of the callsite in
the user's test.

Add an explanation for future readers, explaining callerInfo and when it
should be updated.