fix Test_createPackageMap for 1.15
In Go 1.15 go list will return an error if any of the packages
passed to it are not valid. This differs from pervious behavior
where all of the successful results would be returned. In practice
this should not be an issue as code should contain resolvable
import paths, but this made the test fail.