add -package_format to customize the generated package name
Previously -package must be explictly defined or it will default
to `mock_${name}`. To generate a mock file that sits next to the
package files, eg. as part of the test, one must inspect the package
name first. Which is not scripting friendly.

After this change, we can use -package_format to customize the
generated package name. So that a script can run mockgen without
knowning the package name first.